Created Drink Object and Config interface
I create a drink object that can be inistianted using a JSON object, saved
to a JSON object, or created via a string and a map of ingreients to
volumes.  The drink object also can make its part of the byte array used
to send a 'make drink' command to the arduino.  Also, the drink object can
check, using a list of available ingreidents, if it can be made.

The config interface must be implemented by any config class.  Its
spesifications are discribed within.