Started arduino comunicator class
I started work on the arduino comunicator class with will be responsible
for all the soccket connections to the arduino.  I was careful to make
sure to use timeouts whenever doing a read from the socket to make sure
the program doesn't hang when it sends a request.  Might want to move
requests to a seperate thread.  Currently the com only sends a make drink
command using the 'no-force' flag.

Had to mod config interface to have functionality to get the IP address of
the arduino

chose to mode Drink to return an entire byte array instead of all but the
first byte (the flag).  This is the only command that requires a full
drink spesification, thus it made sense.